+++ src/index.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { lazy, Suspense } from "react";
 import ReactDOM from "react-dom/client";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { ProSidebarProvider } from "react-pro-sidebar";
@@ -8,7 +8,7 @@ import "./assets/styles/bootstrap.scss";
 import App from "./pages/App";
 import Datenschutz from "./pages/Datenschutz";
 import Impressum from "./pages/Impressum";
-import Dashboard from "./pages/Dashboard";
+// import Dashboard from "./pages/Dashboard";
 import Login from "./pages/Login";
 import Registrieren from "./pages/Registrieren";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
@@ -17,11 +17,13 @@ import Zones from "./pages/Zones";
 import { AuthProvider } from "./service/auth";
 import AuthRoute from "./components/routes/AuthRoute";
 import GuestRoute from "./components/routes/GuestRoute";
-import Advanced from "./pages/Advanced";
+// import Advanced from "./pages/Advanced";
 import NotFound from "./pages/NotFound";
 import RoleRoute from "./components/routes/RoleRoute";
 import { ToastContainer } from 'react-toastify';
 
+const Advanced = lazy(() => import("./pages/Advanced"))
+const Dashboard = lazy(() => import("./pages/Dashboard"))
 
 const queryClient = new QueryClient(); // react-query config
 
@@ -39,11 +41,11 @@ const queryClient = new QueryClient(); // react-query config
               <Route path="/register" element={<GuestRoute><Registrieren /></GuestRoute>} />
               <Route path="/" element={<AuthRoute><App /></AuthRoute>}>
                 <Route index element={<Navigate to={"/dashboard"} replace />}></Route>
-                <Route path="dashboard" element={<Dashboard />} />
+                <Route path="dashboard" element={<Suspense fallback=""><Dashboard /></Suspense>} />
                 <Route path="zones" element={<Zones />} />
                 <Route path="map" element={"test"} />
                 <Route path="help" element={"hilfeseite"} />
-                <Route path="advanced" element={<RoleRoute><Advanced /></RoleRoute>} />
+                <Route path="advanced" element={<RoleRoute><Suspense fallback=""><Advanced /></Suspense></RoleRoute>} />
                 <Route path="datenschutz" element={<Datenschutz />} />
                 <Route path="impressum" element={<Impressum />} />
                 <Route path="settings">
